+++ packages/rdl-copilot/src/index.ts
@@ -99,3 +99,4 @@ export const parseNaturalLanguageReportRequest = (
 export * from "./generator";
 export * from "./approved-resources";
 export * from "./xsd-validator";
+export * from "./inspection";